+++ app/src/androidTest/java/com/Poli/pollinatorpathway/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.Poli.home;
+package com.Poli.pollinatorpathway;
 
 import android.content.Context;
