+++ app/src/main/java/com/Poli/pollinatorpathway/MainActivity.java
@@ -1,4 +1,4 @@
-package com.Poli.home;
+package com.Poli.pollinatorpathway;
 
 import android.os.Bundle;
 
@@ -10,7 +10,7 @@ import androidx.navigation.Navigation;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
 
-import com.Poli.home.databinding.ActivityMainBinding;
+import com.Poli.pollinatorpathway.databinding.ActivityMainBinding;
 
 public class MainActivity extends AppCompatActivity {
