+++ app/src/test/java/com/Poli/pollinatorpathway/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.Poli.home;
+package com.Poli.pollinatorpathway;
 
 import org.junit.Test;
